finish Authorization

+++ part4/bloglist/controllers/blogs.js
@@ -1,9 +1,15 @@
 const blogsRouter = require("express").Router();
 const { request, response } = require("../app");
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  });
   response.json(blogs);
 });
 
@@ -13,34 +19,77 @@ blogsRouter.get("/:id", async (request, response) => {
 });
 
 blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body);
+  const body = new Blog(request.body);
 
-  const result = await blog.save();
-  response.status(201).json(result);
+  if (!request.user) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const user = await User.findById(request.user.id);
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id,
+  });
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+  response.status(201).json(savedBlog);
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  if (!request.user) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const user = request.user;
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    response.status(404).json({ error: "Not Find determine Blog" });
+  } else {
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      response.status(401).json({ error: "The blog not belong to User" });
+    }
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
   const body = request.body;
 
-  const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-  };
+  if (!request.user) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
 
-  const updateBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
-    new: true,
-    runValidators: true,
-    context: "query",
-  });
-
-  response.json(updateBlog);
+  const blog = await Blog.findById(request.params.id);
+  const user = request.user;
+  if (!blog) {
+    response.status(404).json({ error: "Not Find determine Blog" });
+  } else {
+    if (blog.user.toString() === user.id.toString()) {
+      const blog = {
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+      };
+      const updateBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
+        new: true,
+        runValidators: true,
+        context: "query",
+      });
+      response.json(updateBlog);
+    } else {
+      response.status(401).json({ error: "The blog not belong to User" });
+    }
+  }
 });
 
 module.exports = blogsRouter;
